commonization CustomCopyToClipBoard component
